feat: add skeleton loading for products
