+++ src/pages/Catalog/components/ProductsSkeleton/ProductsSkeleton.tsx
@@ -1,0 +1,11 @@
+import { SimpleGrid, Skeleton } from '@mantine/core';
+
+export const ProductsSkeleton = () => {
+  return (
+    <SimpleGrid cols={3}>
+      {new Array(6).fill('').map((_, index) => (
+        <Skeleton key={index} h={300} radius="md" width="100%" />
+      ))}
+    </SimpleGrid>
+  );
+};
