+++ src/pages/Catalog/components/ProductsSkeleton/index.ts
@@ -1,0 +1,1 @@
+export * from './ProductsSkeleton';
