+++ src/pages/Catalog/components/index.ts
@@ -2,3 +2,4 @@ export * from './ProductCard';
 export * from './SidebarFilters';
 export * from './TreeCategoryLink';
 export * from './BreadcrumbsCategories';
+export * from './ProductsSkeleton';
